+++ src/docker.rs
@@ -45,12 +45,12 @@ pub fn run(target: Target,
     fs::create_dir(&cargo_dir).ok();
     fs::create_dir(&xargo_dir).ok();
 
-    let mut cmd = if target.uses_xargo() {
+    let mut cargo = if target.uses_xargo() {
         Command::new("xargo")
     } else {
         Command::new("cargo")
     };
-    cmd.args(args);
+    cargo.args(args);
 
     let version = env!("CARGO_PKG_VERSION");
     let tag = if version.ends_with("-dev") {
@@ -59,12 +59,14 @@ pub fn run(target: Target,
         Cow::from(format!("v{}", version))
     };
 
-    Command::new("docker")
-        .arg("run")
+    let mut docker = Command::new("docker");
+
+    docker.arg("run")
         .arg("--rm")
         .args(&["--user", &format!("{}:{}", id::user(), id::group())])
         .args(&["-e", "CARGO_HOME=/cargo"])
         .args(&["-e", "CARGO_TARGET_DIR=/target"])
+        .args(&["-e", "HOME=/tmp"])
         .args(&["-e", &format!("USER={}", id::username())])
         .args(&["-e", "XARGO_HOME=/xargo"])
         .args(&["-v", &format!("{}:/xargo", xargo_dir.display())])
@@ -72,9 +74,45 @@ pub fn run(target: Target,
         .args(&["-v", &format!("{}:/project:ro", cargo_root.display())])
         .args(&["-v",
                 &format!("{}:/rust:ro", rustc::sysroot(verbose)?.display())])
-        .args(&["-v", &format!("{}:/target", target_dir.display())])
-        .args(&["-w", "/project"])
-        .args(&["-it", &format!("japaric/{}:{}", target.triple(), tag)])
-        .args(&["sh", "-c", &format!("PATH=$PATH:/rust/bin {:?}", cmd)])
-        .run_and_get_status(verbose)
+        .args(&["-v", &format!("{}:/target", target_dir.display())]);
+
+    // `emcc` caches stuff in `$HOME`. We must preserve that information across
+    // runs so we use an OS-level temporary directory for that.
+    let must_run_emcc_first = if target.uses_emcc() {
+        let temp_dir = env::temp_dir();
+        let cache_dir = temp_dir.join(format!("cross-{}", target.triple()));
+
+        if !cache_dir.exists() {
+            fs::create_dir(&cache_dir).chain_err(|| {
+                    format!("couldn't create a directory in {}",
+                            temp_dir.display())
+                })?;
+        }
+
+        docker.args(&["-v", &format!("{}:{}", cache_dir.display(), "/tmp")]);
+
+        !cache_dir.join(".emscripten").exists()
+    } else {
+        false
+    };
+
+    docker.args(&["-w", "/project"])
+        .args(&["-it", &format!("japaric/{}:{}", target.triple(), tag)]);
+
+    // `emcc` only creates the `~/.emscripten` file the very first time is
+    // called. Make sure that first time is from "outside" `rustc` so that
+    // `cross build` doesn't result in a no-op.
+    if must_run_emcc_first {
+        docker.args(&["sh",
+                      "-c",
+                      &format!("emcc 2>/dev/null; \
+                                PATH=$PATH:/rust/bin {:?}",
+                               cargo)]);
+    } else {
+        docker.args(&["sh",
+                      "-c",
+                      &format!("PATH=$PATH:/rust/bin {:?}", cargo)]);
+    }
+
+    docker.run_and_get_status(verbose)
 }